requirements.txt
# Application RAG avec Azure AI Search et chiffrement
flask==3.0.3
flask-cors==4.0.0
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.8.0
azure-keyvault-keys==4.9.0
azure-identity==1.16.1
azure-search-documents==11.4.0
azure-ai-documentintelligence==1.0.0
azure-core>=1.30.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
Pillow==10.4.0
requests==2.32.3
cryptography>=42.0.0
gunicorn==21.2.0
Werkzeug==3.0.3

# Packages pour le chiffrement IronCore CloakedAI
ironcore-alloy==0.10.3  # Package non disponible publiquement
